changed folders names and moved ResultBox.js
